% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{BGDR}
\alias{BGDR}
\title{A function to compute Coefficients of the unary linear regression equation by Batch Gradient Descent approach with fixed step}
\usage{
BGDR(x, y, theta, step, maxiter = 10000, epsilon = 1e-04)
}
\arguments{
\item{x}{the covariate}

\item{y}{the response variable}

\item{theta}{the initial value of beta}

\item{step}{the learning rate}

\item{maxiter}{the max number of iteration}

\item{epsilon}{the end condition}
}
\value{
the estimation
}
\description{
A function to compute Coefficients of the unary linear regression equation by Batch Gradient Descent approach with fixed step
}
\examples{
\dontrun{
n <- 5000
x <- runif(n, 0, 2)
y <- x * 4. + 5. + runif(n)
BGD(x, y, theta, 0.1)
}
}
