% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GibbsC}
\alias{GibbsC}
\title{A Gibbs sampler using Rcpp}
\usage{
GibbsC(N, x0, n, a, b)
}
\arguments{
\item{N}{the number of samples}

\item{x0}{Initial value}

\item{n}{A given value}

\item{a}{A given value}

\item{b}{A given value}
}
\value{
a random sample of size \code{n}
}
\description{
A Gibbs sampler using Rcpp
}
\examples{
\dontrun{
Y<-GibbsC(N, x0, n, a, b)
plot(Y[1001:N,1],Y[1001:N,2],xlab = "y1",ylab = "y2")
}
}
