% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{MBGDR}
\alias{MBGDR}
\title{A function to compute Coefficients of the unary linear regression equation by Mini-batch Gradient Descent approach with fixed step}
\usage{
MBGDR(x, y, theta, step, batch = 10, maxiter = 10000, epsilon = 1e-04)
}
\arguments{
\item{x}{the covariate}

\item{y}{the response variable}

\item{theta}{the initial value of beta}

\item{step}{the learning rate}

\item{batch}{the batch size}

\item{maxiter}{the max number of iteration}

\item{epsilon}{the end condition}
}
\value{
the estimation
}
\description{
A function to compute Coefficients of the unary linear regression equation by Mini-batch Gradient Descent approach with fixed step
}
\examples{
\dontrun{
MBGD(x, y, theta, 0.1)
}
}
