\name{StatComp21068-package}
\alias{StatComp21068-package}
\alias{StatComp21068}
\docType{package}
\title{
  This package provides several functions for 'Statistical Computing' course
}
\description{
  This package provides three functions to compute Coefficients of the unary linear regression equation by Mini-batch Gradient Descent approach with fixed step
}
\author{
Junfeng Zhao

Maintainer: Junfeng Zhao <2215235986@qq.com>
}
\keyword{ package }
